+++ sketch-extensions/src/androidTest/java/com/github/panpf/sketch/extensions/test/SketchImageViewTest.kt
@@ -30,8 +30,8 @@ import com.github.panpf.sketch.request.DisplayResult.Error
 import com.github.panpf.sketch.request.DisplayResult.Success
 import com.github.panpf.sketch.request.ImageOptions
 import com.github.panpf.sketch.request.Listener
-import com.github.panpf.sketch.request.internal.Listeners
 import com.github.panpf.sketch.request.ProgressListener
+import com.github.panpf.sketch.request.internal.Listeners
 import com.github.panpf.sketch.request.internal.ProgressListeners
 import com.github.panpf.sketch.resize.Precision.SAME_ASPECT_RATIO
 import com.github.panpf.sketch.resize.Scale.END_CROP
@@ -86,6 +86,7 @@ class SketchImageViewTest {
         (LayoutInflater.from(context)
             .inflate(R.layout.attrs_test_state, null, false) as SketchImageView).apply {
             Assert.assertNotNull(displayImageOptions!!.placeholder)
+            Assert.assertNotNull(displayImageOptions!!.uriEmpty)
             Assert.assertNotNull(displayImageOptions!!.error)
             Assert.assertNotNull((displayImageOptions!!.error as ErrorStateImage).stateList.find { it.first is UriEmptyCondition })
         }