+++ sketch/src/androidTest/java/com/github/panpf/sketch/test/request/DisplayRequestTest.kt
@@ -148,6 +148,7 @@ class DisplayRequestTest {
             Assert.assertFalse(this.ignoreExifOrientation)
             Assert.assertEquals(ENABLED, this.resultCachePolicy)
             Assert.assertNull(this.placeholder)
+            Assert.assertNull(this.uriEmpty)
             Assert.assertNull(this.error)
             Assert.assertNull(this.transitionFactory)
             Assert.assertFalse(this.disallowAnimatedImage)
@@ -185,6 +186,7 @@ class DisplayRequestTest {
             Assert.assertFalse(this.ignoreExifOrientation)
             Assert.assertEquals(ENABLED, this.resultCachePolicy)
             Assert.assertNull(this.placeholder)
+            Assert.assertNull(this.uriEmpty)
             Assert.assertNull(this.error)
             Assert.assertNull(this.transitionFactory)
             Assert.assertFalse(this.disallowAnimatedImage)
@@ -1320,6 +1322,40 @@ class DisplayRequestTest {
             placeholder(null)
             build().apply {
                 Assert.assertNull(placeholder)
+            }
+        }
+    }
+
+    @Test
+    fun testUriEmpty() {
+        val context1 = getTestContext()
+        val uriString1 = newAssetUri("sample.jpeg")
+        DisplayRequest.Builder(context1, uriString1).apply {
+            build().apply {
+                Assert.assertNull(uriEmpty)
+            }
+
+            uriEmpty(ColorStateImage(IntColor(Color.BLUE)))
+            build().apply {
+                Assert.assertEquals(ColorStateImage(IntColor(Color.BLUE)), uriEmpty)
+            }
+
+            uriEmpty(ColorDrawable(Color.GREEN))
+            build().apply {
+                Assert.assertEquals(true, uriEmpty is DrawableStateImage)
+            }
+
+            uriEmpty(drawable.bottom_bar)
+            build().apply {
+                Assert.assertEquals(
+                    DrawableStateImage(drawable.bottom_bar),
+                    uriEmpty
+                )
+            }
+
+            uriEmpty(null)
+            build().apply {
+                Assert.assertNull(uriEmpty)
             }
         }
     }