+++ sketch/src/main/java/com/github/panpf/sketch/request/LoadRequest.kt
@@ -359,6 +359,18 @@ interface LoadRequest : ImageRequest {
             super.placeholder(drawableResId)
         }
 
+        override fun uriEmpty(stateImage: StateImage?): Builder = apply {
+            super.uriEmpty(stateImage)
+        }
+
+        override fun uriEmpty(drawable: Drawable): Builder = apply {
+            super.uriEmpty(drawable)
+        }
+
+        override fun uriEmpty(drawableResId: Int): Builder = apply {
+            super.uriEmpty(drawableResId)
+        }
+
         override fun error(
             defaultStateImage: StateImage?, configBlock: (ErrorStateImage.Builder.() -> Unit)?
         ): Builder = apply {
@@ -454,6 +466,7 @@ interface LoadRequest : ImageRequest {
         override val ignoreExifOrientation: Boolean,
         override val resultCachePolicy: CachePolicy,
         override val placeholder: StateImage?,
+        override val uriEmpty: StateImage?,
         override val error: ErrorStateImage?,
         override val transitionFactory: Factory?,
         override val disallowAnimatedImage: Boolean,