+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * change: No longer intercept requests with 'Transfer-Encoding' is 'chunked'
 * fix: Fixed a bug where other images may be displayed unexpectedly when uri is empty
 * improve: Built-in exceptions no longer print stack information
+* new: ImageRequest and ImageOptions add 'uriEmpty()' method for more convenient configuration of uri empty state image
 
 #### sketch-zoom: